feature/lab5 basic 1 fin

+++ lab5/lib/include/timer.h
@@ -1,3 +1,5 @@
+#include "stdint.h"
+
 #define CORE0_TIMER_IRQ_CTRL 0x40000040
 
 #define STR(x) #x
@@ -12,7 +14,7 @@ unsigned long clock_time(void);
 void clock_alert(char *str);
 void timeout_print(char *str);
 void print_time(void);
-void set_core_timer_interrupt(unsigned long long expired_time);
+void set_core_timer_interrupt(uint64_t expired_time);
 void add_timer(callback_typ callback, char *msg, int time);
 void pop_timer(void);
 void add_task(callback_typ callback, char *msg, int piority);
@@ -21,7 +23,7 @@ void pop_task(void);
 typedef struct timer_list timer_list;
 
 struct timer_list{
-  unsigned long long expired_time;
+  uint64_t expired_time;
   callback_typ call_back;
   char msg[100];
   struct timer_list *next;
